feat(note): show full created date format on selected notes

+++ crates/notedeck/src/time.rs
@@ -84,9 +84,9 @@ fn time_ago_between(i18n: &mut Localization, timestamp: u64, now: u64) -> String
     }
 }
 
-pub fn time_format(_i18n: &mut Localization, timestamp: i64) -> String {
+pub fn time_format(_i18n: &mut Localization, timestamp: u64) -> String {
     // TODO: format this using the selected locale
-    DateTime::from_timestamp(timestamp, 0)
+    DateTime::from_timestamp(timestamp as i64, 0)
         .unwrap()
         .format("%l:%M %p %b %d, %Y")
         .to_string()
